No Norm and int16 for data
Removed Normalization and forced the image data to be of the data type int16

+++ GZimp/Package.py
@@ -62,11 +62,12 @@ class Images():
         del i
 
 
-        images = np.zeros(mult).reshape(shape) 
+        images = np.zeros(mult).reshape(shape).astype('int16') 
 
         for i in range(len(sample)):
 
             img=  image.imread(sample[i])
+            img = img.astype('int16') 
             img = img[crop : -1*crop , crop : -1*crop]
             #img = cv2.normalize(img, None, 0, 1.0, cv2.NORM_MINMAX, dtype=cv2.CV_32F)
             img = np.expand_dims(img, axis=0)
